+++ frameworks/Go/fiber/fiber.dockerfile
@@ -1,12 +1,10 @@
 FROM golang:1.13
 
-ENV GO111MODULE on
 WORKDIR /fiber
 
 COPY ./src /fiber
 
 RUN go mod download
-
-RUN go build -o server server.go
+RUN go build -o server *.go
 
 CMD ./server
