+++ frameworks/Go/fiber/src/go.mod
@@ -3,7 +3,8 @@ module fiber
 go 1.13
 
 require (
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/gofiber/fiber v1.2.0
-	github.com/jackc/pgx/v4 v4.2.1
+	github.com/gofiber/fiber v1.2.3
+	github.com/jackc/pgx/v4 v4.3.0
+	github.com/tidwall/gjson v1.4.0 // indirect
+	github.com/tidwall/sjson v1.0.4
 )
